+++ tests/fixtures/helpers.py
@@ -1,0 +1,23 @@
+"""Helpers fixture for ``movement`` test modules."""
+
+import pytest
+
+
+class Helpers:
+    """General helper methods for ``movement`` test modules."""
+
+    @staticmethod
+    def count_nans(da):
+        """Count number of NaNs in a DataArray."""
+        return da.isnull().sum().item()
+
+    @staticmethod
+    def count_consecutive_nans(da):
+        """Count occurrences of consecutive NaNs in a DataArray."""
+        return (da.isnull().astype(int).diff("time") != 0).sum().item()
+
+
+@pytest.fixture
+def helpers():
+    """Return an instance of the ``Helpers`` class."""
+    return Helpers
